Fixed situation where callback is undefined and was being executed
Callback was only being checked to be a function in one location, not
everytime it was used. Instead callback is now checked to be a function
at the top otherwise a empty function is assigned to it so the
executions don’t cause errors.

```
/node_modules/fancy-errors/lib/index.js:64
      callback(err);
      ^
TypeError: undefined is not a function
    at Object.catchErr [as catch]
(/node_modules/fancy-errors/lib/index.js:64:7)
```